+++ packages/frontend/core/src/blocksuite/ai/chat-panel/chat-context.ts
@@ -1,5 +1,3 @@
-import type { Signal } from '@preact/signals-core';
-
 import type { AIError } from '../provider';
 
 export type ChatMessage = {
@@ -34,22 +32,6 @@ export type ChatStatus =
   | 'idle'
   | 'transmitting';
 
-export interface DocContext {
-  docId: string;
-  docTitle: string;
-  docContent: string;
-  tags: string;
-  createDate: string;
-  updatedDate: string;
-}
-
-export interface FileContext {
-  blobId: string;
-  fileName: string;
-  fileType: string;
-  fileContent: string;
-}
-
 export type ChatContextValue = {
   // history messages of the chat
   items: ChatItem[];
@@ -61,10 +43,6 @@ export type ChatContextValue = {
   markdown: string;
   // images of the selected content or user uploaded
   images: File[];
-  // chips of workspace doc or user uploaded file
-  chips: ChatChip[];
-  // the progress of the embedding
-  embeddingProgress: [number, number];
   abortController: AbortController | null;
 };
 
@@ -73,39 +51,3 @@ export type ChatBlockMessage = ChatMessage & {
   userName?: string;
   avatarUrl?: string;
 };
-
-export type ChipState = 'candidate' | 'processing' | 'finished' | 'failed';
-
-export interface BaseChip {
-  /**
-   * candidate: the chip is a candidate for the chat
-   * processing: the chip is processing
-   * finished: the chip is successfully processed
-   * failed: the chip is failed to process
-   */
-  state: ChipState;
-  tooltip?: string | null;
-  createdAt?: number | null;
-}
-
-export interface DocChip extends BaseChip {
-  docId: string;
-  markdown?: Signal<string> | null;
-  tokenCount?: number | null;
-}
-
-export interface FileChip extends BaseChip {
-  file: File;
-  fileId?: string | null;
-  blobId?: string | null;
-}
-
-export interface TagChip extends BaseChip {
-  tagId: string;
-}
-
-export interface CollectionChip extends BaseChip {
-  collectionId: string;
-}
-
-export type ChatChip = DocChip | FileChip | TagChip | CollectionChip;
