+++ packages/frontend/core/src/blocksuite/ai/components/ai-chat-chips/file-chip.ts
@@ -4,7 +4,7 @@ import { ShadowlessElement } from '@blocksuite/affine/std';
 import { html } from 'lit';
 import { property } from 'lit/decorators.js';
 
-import type { FileChip } from '../chat-context';
+import type { FileChip } from './type';
 import { getChipIcon, getChipTooltip } from './utils';
 
 export class ChatPanelFileChip extends SignalWatcher(
