+++ packages/frontend/core/src/blocksuite/ai/components/ai-chat-chips/index.ts
@@ -1,0 +1,2 @@
+export * from './type';
+export * from './utils';
