+++ packages/frontend/core/src/blocksuite/ai/effects.ts
@@ -23,16 +23,8 @@ import { ActionMindmap } from './chat-panel/actions/mindmap';
 import { ActionSlides } from './chat-panel/actions/slides';
 import { ActionText } from './chat-panel/actions/text';
 import { AILoading } from './chat-panel/ai-loading';
-import { ChatPanelChips } from './chat-panel/chat-panel-chips';
 import { ChatPanelInput } from './chat-panel/chat-panel-input';
 import { ChatPanelMessages } from './chat-panel/chat-panel-messages';
-import { ChatPanelAddPopover } from './chat-panel/components/add-popover';
-import { ChatPanelCandidatesPopover } from './chat-panel/components/candidates-popover';
-import { ChatPanelChip } from './chat-panel/components/chip';
-import { ChatPanelCollectionChip } from './chat-panel/components/collection-chip';
-import { ChatPanelDocChip } from './chat-panel/components/doc-chip';
-import { ChatPanelFileChip } from './chat-panel/components/file-chip';
-import { ChatPanelTagChip } from './chat-panel/components/tag-chip';
 import { AssistantAvatar } from './chat-panel/content/assistant-avatar';
 import { ChatContentImages } from './chat-panel/content/images';
 import { ChatContentPureText } from './chat-panel/content/pure-text';
@@ -40,6 +32,14 @@ import { ChatContentRichText } from './chat-panel/content/rich-text';
 import { ChatMessageAction } from './chat-panel/message/action';
 import { ChatMessageAssistant } from './chat-panel/message/assistant';
 import { ChatMessageUser } from './chat-panel/message/user';
+import { ChatPanelAddPopover } from './components/ai-chat-chips/add-popover';
+import { ChatPanelCandidatesPopover } from './components/ai-chat-chips/candidates-popover';
+import { ChatPanelChips } from './components/ai-chat-chips/chat-panel-chips';
+import { ChatPanelChip } from './components/ai-chat-chips/chip';
+import { ChatPanelCollectionChip } from './components/ai-chat-chips/collection-chip';
+import { ChatPanelDocChip } from './components/ai-chat-chips/doc-chip';
+import { ChatPanelFileChip } from './components/ai-chat-chips/file-chip';
+import { ChatPanelTagChip } from './components/ai-chat-chips/tag-chip';
 import { effects as componentAiItemEffects } from './components/ai-item';
 import { AIScrollableTextRenderer } from './components/ai-scrollable-text-renderer';
 import { AskAIButton } from './components/ask-ai-button';
